pyproject.toml
[project]
name = "scm-whatsapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.6.3",
    "fastapi>=0.115.13",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "litellm>=1.73.0",
    "pgvector>=0.4.1",
    "psycopg>=3.2.9",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "twilio>=9.6.3",
    "uvicorn>=0.34.3",
    "xmltodict>=0.14.2",
]
